// src/idea/idea.dto.ts
/* eslint-disable prettier/prettier */
import { IsString } from 'class-validator';
import { UserRO } from 'src/user/user.dto';

export class IdeaDTO {
    @IsString()
    readonly idea: string;
  
    @IsString()
    readonly description: string;
}

export class IdeaRO {
    id: string;
    created: Date;
    updated: Date;
    idea: string;
    description: string;
    author: UserRO;
    upvotes?: number;
  downvotes?: number;
}